can now calculate the basic trig identities

+++ src/main/java/Rules/PowerRule.java
@@ -27,7 +27,7 @@ public class PowerRule extends DerivationRule {
 
     @Override
     public double getResult(List<Integer> dims) {
-        return (int)Math.pow((double)this.terms.get(0).evaluate(dims), (double)this.terms.get(1).evaluate(dims));
+        return Math.pow(this.terms.get(0).evaluate(dims), this.terms.get(1).evaluate(dims));
     }
 
 }
